:rocket: Add new package to example project
