+++ nexpo-example/package.json
@@ -9,7 +9,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-native-web": "^0.18.10",
-    "rn-responsive-styles": "^2.0.1"
+    "rn-responsive-styles": "^2.1.0"
   },
   "devDependencies": {
     "babel-plugin-react-native-web": "^0.18.10",
