+++ nexpo-example/pages/index.tsx
@@ -1,34 +1,11 @@
-import { Text, View } from 'react-native'
-import { CreateResponsiveStyle, DEVICE_SIZES, maxSize } from 'rn-responsive-styles'
+import { ActivityIndicator } from 'react-native'
+import Home from './home'
+import { SSRProvider } from '../../src'
 
 export default function App() {
-  const styles = useStyles()
-  console.log(styles.link)
   return (
-    <View>
-      <Text
-        style={styles.link}
-        accessibilityRole="link"
-        //@ts-ignore
-        href={`/alternate`}
-      >
-        A universal link
-      </Text>
-    </View>
+    <SSRProvider placeholder={<ActivityIndicator />}>
+      <Home />
+    </SSRProvider>
   )
 }
-
-const useStyles = CreateResponsiveStyle(
-  {
-    link: {
-      color: 'blue',
-    },
-  },
-  {
-    [maxSize(DEVICE_SIZES.SM)]: {
-      link: {
-        color: 'red',
-      },
-    },
-  }
-)
